Refactor Package info into metadata module
